HPCC-27294 Enable hthor+roxie DFS service access
